feat: move request to api
